Move various components and function definitions
