+++ src/app/components/content/IsaacQuestion.tsx
@@ -1,9 +1,7 @@
-import React, {ContextType, Suspense, useContext, useEffect} from "react";
+import React, {Suspense, useContext, useEffect} from "react";
 import {
-    attemptQuestion,
     deregisterQuestions,
     registerQuestions,
-    saveGameboard,
     selectors,
     useAppDispatch,
     useAppSelector
@@ -18,20 +16,14 @@ import {
     determineFastTrackSecondaryAction,
     fastTrackProgressEnabledBoards,
     isAda,
-    isLoggedIn,
-    isNotPartiallyLoggedIn,
     isPhy,
-    KEY,
-    persistence,
     QUESTION_TYPES,
     selectQuestionPart,
-    trackEvent,
+    submitCurrentAttempt,
     useDeviceSize,
-    useFastTrackInformation,
-    wasTodayUTC
-} from "../../services";
+    useFastTrackInformation} from "../../services";
 import {DateString, TIME_ONLY} from "../elements/DateString";
-import {AccordionSectionContext, AppQuestionDTO, ConfidenceContext, GameboardContext, InlineQuestionDTO, InlineContext} from "../../../IsaacAppTypes";
+import {AccordionSectionContext, ConfidenceContext, GameboardContext, InlineQuestionDTO, InlineContext} from "../../../IsaacAppTypes";
 import {RouteComponentProps, withRouter} from "react-router";
 import {IsaacLinkHints, IsaacTabbedHints} from "./IsaacHints";
 import {ConfidenceQuestions, useConfidenceQuestionsValues} from "../elements/inputs/ConfidenceQuestions";
@@ -105,7 +97,6 @@ export const IsaacQuestion = withRouter(({doc, location}: {doc: ApiTypes.Questio
     const inlineContext = useContext(InlineContext);
     const isInlineQuestion = doc.type === "isaacInlineRegion" && inlineContext;
 
-    const inlineElementIds = Object.keys(inlineContext?.elementToQuestionMap ?? {});
     const numInlineQuestions = isInlineQuestion ? Object.values(inlineContext?.elementToQuestionMap ?? {}).length : undefined;
     const numCorrectInlineQuestions = (isInlineQuestion && validationResponse) ? (questionPart as InlineQuestionDTO).validationResponse?.partsCorrect : undefined;
     const showInlineAttemptStatus = !isInlineQuestion || !inlineContext?.isModifiedSinceLastSubmission;
@@ -245,22 +236,3 @@ export const IsaacQuestion = withRouter(({doc, location}: {doc: ApiTypes.Questio
         </RS.Form>
     </ConfidenceContext.Provider>;
 });
-
-export const submitCurrentAttempt = (questionPart: AppQuestionDTO | undefined, docId: string, currentGameboard: ApiTypes.GameboardDTO | undefined, currentUser: any, dispatch: any, inlineContext?: ContextType<typeof InlineContext>) => {
-    if (questionPart?.currentAttempt) {
-        // Notify Plausible that at least one question attempt has taken place today
-        if (persistence.load(KEY.INITIAL_DAILY_QUESTION_ATTEMPT_TIME) == null || !wasTodayUTC(persistence.load(KEY.INITIAL_DAILY_QUESTION_ATTEMPT_TIME))) {
-            persistence.save(KEY.INITIAL_DAILY_QUESTION_ATTEMPT_TIME, new Date().toString());
-            trackEvent("question_attempted");
-        }
-
-        dispatch(attemptQuestion(docId, questionPart?.currentAttempt, currentGameboard?.id, inlineContext));
-        if (isLoggedIn(currentUser) && isNotPartiallyLoggedIn(currentUser) && currentGameboard?.id && !currentGameboard.savedToCurrentUser) {
-            dispatch(saveGameboard({
-                boardId: currentGameboard.id,
-                user: currentUser,
-                redirectOnSuccess: false
-            }));
-        }
-    }
-};
